# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Api QuickMenu",
        "description": "Documentação da API do sistema de restaurante quickmenu.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "Servidor principal"
        }
    ],
    "paths": {
        "/api/addresses": {
            "get": {
                "tags": [
                    "Address"
                ],
                "summary": "List all addresses",
                "description": "Retrieve list of all addresses",
                "operationId": "d36bf2c013f585f04a4b79829ac21883",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "addresses": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Address"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Address"
                ],
                "summary": "Create address",
                "description": "Store address",
                "operationId": "6bbc1d4e6af4b09f0e76bc48dc84ea95",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Address"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Endereço registrado com sucesso"
                    }
                }
            }
        },
        "/api/addresses/{address}": {
            "get": {
                "tags": [
                    "Address"
                ],
                "summary": "Get address by id",
                "description": "Retrieve a single address",
                "operationId": "2afc0f801cfd83bca228c2c15ec9a5f8",
                "parameters": [
                    {
                        "name": "address",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Address"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Address"
                ],
                "summary": "Update address",
                "description": "Update address data",
                "operationId": "03c1acda1e9e3f86f3f388edfb0f1e33",
                "parameters": [
                    {
                        "name": "address",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Address"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Endereço atualizado com sucesso"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Address"
                ],
                "summary": "Delete address",
                "description": "Permanently delete address",
                "operationId": "a9355aee849aeb0e9e73b2c42b77ea39",
                "parameters": [
                    {
                        "name": "address",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Endereço deletado com sucesso"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "Address": {
                "title": "Address",
                "required": [
                    "address_name",
                    "quatrain",
                    "number",
                    "city",
                    "state",
                    "cep"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "readOnly": true
                    },
                    "address_name": {
                        "description": "Nome do endereço",
                        "type": "string"
                    },
                    "quatrain": {
                        "description": "Quadra",
                        "type": "string"
                    },
                    "number": {
                        "type": "string"
                    },
                    "complement": {
                        "type": "string",
                        "nullable": true
                    },
                    "district": {
                        "type": "string",
                        "nullable": true
                    },
                    "city": {
                        "type": "string"
                    },
                    "state": {
                        "type": "string"
                    },
                    "neighborhood": {
                        "type": "string",
                        "nullable": true
                    },
                    "cep": {
                        "type": "string"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "readOnly": true
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "readOnly": true
                    }
                },
                "type": "object"
            }
        }
    },
...
}